# openapi.yaml
openapi: 3.0.3
info:
  title: Nutri Thata API
  description: API para plataforma de educação nutricional com sistema de módulos, convites e upload de mídia
  version: 1.0.0
  contact:
    name: API Support
    email: support@nutrithata.com
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://api.nutrithata.com
    description: Produção
security:
  - BearerAuth: []
paths:
...
  /invites:
    post:
      tags:
        - Convites
      summary: Criar convite para paciente
      description: Envia convite por email para novo paciente (PROFESSIONAL ou ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: "paciente@email.com"
                prefill:
                  type: object
                  description: Dados para pré-preenchimento do formulário
                  properties:
                    name:
                      type: string
                      example: "Maria Santos"
                    phone:
                      type: string
                      example: "(11) 99999-9999"
      responses:
        '201':
          description: Convite criado e enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Convite enviado com sucesso"
                  data:
                    type: object
                    properties:
                      inviteId:
                        type: string
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      email:
                        type: string
                        example: "paciente@email.com"
                      expiresAt:
                        type: string
                        format: date-time
                        example: "2024-02-15T23:59:59Z"
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão (apenas PROFESSIONAL ou ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email já possui convite pendente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Convites
      summary: Listar convites criados
      description: Lista convites criados pelo profissional autenticado
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por status do convite
          schema:
            type: string
            enum: ["PENDING", "ACCEPTED", "EXPIRED", "CANCELLED"]
        - name: page
          in: query
          description: Número da página (base 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista de convites
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      invites:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invite'
                      totalElements:
                        type: integer
                        example: 15
                      totalPages:
                        type: integer
                        example: 2
                      currentPage:
                        type: integer
                        example: 0
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{token}:
    get:
      tags:
        - Convites
      summary: Visualizar convite por token
      description: Obtém dados do convite para pré-preenchimento (acesso público)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Token único do convite
          schema:
            type: string
            example: "abc123def456"
      responses:
        '200':
          description: Dados do convite
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: "paciente@email.com"
                      prefill:
                        type: object
                        properties:
                          name:
                            type: string
                            example: "Maria Santos"
                          phone:
                            type: string
                            example: "(11) 99999-9999"
                      isValid:
                        type: boolean
                        example: true
                      expiresAt:
                        type: string
                        format: date-time
                        example: "2024-02-15T23:59:59Z"
        '404':
          description: Token inválido ou não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Convite expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{token}/accept:
    post:
      tags:
        - Convites
      summary: Aceitar convite e criar conta
      description: Cria conta de paciente baseada no convite (acesso público)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Token único do convite
          schema:
            type: string
            example: "abc123def456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - password
              properties:
                name:
                  type: string
                  example: "Maria Santos"
                password:
                  type: string
                  minLength: 8
                  example: "MinhaSenh@123"
                phone:
                  type: string
                  example: "(11) 99999-9999"
      responses:
        '201':
          description: Conta criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '404':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Convite já foi aceito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Convite expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{id}/resend:
    post:
      tags:
        - Convites
      summary: Reenviar convite
      description: Reenvia email de convite (criador ou ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do convite
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Convite reenviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Convite reenviado com sucesso"
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Convite não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Convite já foi aceito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /invites/{id}/cancel:
    post:
      tags:
        - Convites
      summary: Cancelar convite
      description: Cancela convite pendente (criador ou ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do convite
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Convite cancelado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Convite cancelado com sucesso"
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Convite não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Convite já foi aceito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== RECURSOS ESTÁTICOS ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Login realizado com sucesso"
        data:
          type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiJ9..."
            user:
              $ref: '#/components/schemas/User'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        faultstring:
          type: string
          example: "Credenciais inválidas"
        stacktrace:
          type: string
          description: "Disponível apenas em ambiente de desenvolvimento"
          example: "com.nutrithata.exception.AuthenticationException: Invalid credentials..."

    # ========== MODELOS DE USUÁRIO ==========
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao@email.com"
        role:
          type: string
          enum: ["PATIENT", "PROFESSIONAL", "ADMIN"]
          example: "PATIENT"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Invite:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "paciente@email.com"
        token:
          type: string
          example: "abc123def456ghi789"
        role:
          type: string
          enum: ["PATIENT"]
          example: "PATIENT"
        prefill:
          type: object
          properties:
            name:
              type: string
              example: "Maria Santos"
            phone:
              type: string
              example: "(11) 99999-9999"
        status:
          type: string
          enum: ["PENDING", "ACCEPTED", "EXPIRED", "CANCELLED"]
          example: "PENDING"
        expiresAt:
          type: string
          format: date-time
          example: "2024-02-15T23:59:59Z"
        acceptedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        createdBy:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: "Dra. Ana Paula Silva"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...